Improve website github sync
